requirements.txt
Django==1.11.1
django-debug-toolbar==1.8
django-environ==0.4.4
djangorestframework==3.6.2
psycopg2==2.7.4
pytz==2018.3
six==1.11.0
sqlparse==0.2.4
